+++ changelog.md
@@ -4,8 +4,13 @@
 
 ### Added
 * Support `page-range:*` to select changed pages when using `watch`
+* Add `spread-percent`
+* Add shortcuts for the values of `distribute-rows`, `distribute-columns`, and `distribute-items` 
  
 ### Changed
+* Default `watch` to true, but do not watch if there is nothing to watch
+* Handle multiple books by merging into one
+
 ### Deprecated
 ### Removed
 ### Fixed